refactor(EntityNotFoundException): create exception structure
- 효율적인 에러 처리를 위해 엔티티 미발견에 대한 에러 구조 추가
