+++ src/main/java/nextstep/subway/exceptions/EntityNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.station.application.exceptions;
+package nextstep.subway.exceptions;
 
 public class EntityNotFoundException extends RuntimeException {
     public EntityNotFoundException(final String message) {
