+++ src/main/java/nextstep/subway/line/ui/LineControllerAdvice.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.ui;
 
-import nextstep.subway.station.application.exceptions.EntityNotFoundException;
+import nextstep.subway.exceptions.EntityNotFoundException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
