+++ src/main/java/nextstep/subway/station/application/StationService.java
@@ -1,6 +1,6 @@
 package nextstep.subway.station.application;
 
-import nextstep.subway.station.application.exceptions.EntityNotFoundException;
+import nextstep.subway.exceptions.EntityNotFoundException;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.StationRepository;
 import nextstep.subway.station.dto.StationRequest;
